% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pointerString}
\alias{pointerString}
\title{Character String for the Pointer to an R object}
\usage{
pointerString(x)
}
\arguments{
\item{x}{Any R object}
}
\description{
Returns a character string containing the hexadecimal form of
the \dQuote{address} of an arbitrary R object.
}
\author{
John M. Chambers
}
\keyword{Programming}

