\name{XRtools-package}
\alias{XRtools-package}
\alias{XRtools}
\docType{package}
\title{
\packageTitle{XRtools}
}
\description{
\packageDescription{XRtools}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{XRtools}
\packageIndices{XRtools}

}
\author{
\packageAuthor{XRtools}

Maintainer: \packageMaintainer{XRtools}
}
\keyword{ package }
\keyword{ programming }
